add event bloc and complate data layer

+++ lib/movies/data/models/movie_model.dart
@@ -28,6 +28,7 @@ class MovieModel extends Movie {
         posterPath: json['poster_path'] ?? '',
         releaseDate: json['release_date'] ?? '',
         title: json['title'] ?? '',
-        voteAverage: json['vote_average']);
+        voteAverage: json['vote_average'].toDouble()
+    );
   }
 }
